Raymond_work/Project_Framework/deepwater_inference/requirements.txt
_libgcc_mutex==0.1==main
_tflow_select==2.1.0==gpu
absl-py==0.9.0==py37_0
astor==0.8.0==py37_0
blas==1.0==mkl
blinker==1.4==py37_0
bzip2==1.0.8==h516909a_2
c-ares==1.15.0==h7b6447c_1001
ca-certificates==2020.4.5.1==hecc5488_0
cachetools==3.1.1==py_0
cairo==1.14.12==h8948797_3
certifi==2020.4.5.1==py37hc8dfbb8_0
cffi==1.14.0==py37he30daa8_1
chardet==3.0.4==py37_1003
click==7.1.2==py_0
cloudpickle==1.4.1==py_0
cryptography==2.9.2==py37h1ba5d50_0
cudatoolkit==10.1.243==h6bb024c_0
cudnn==7.6.5==cuda10.1_0
cupti==10.1.168==0
cycler==0.10.0==py37_0
cytoolz==0.10.1==py37h7b6447c_0
dask-core==2.16.0==py_0
dbus==1.13.14==hb2f20db_0
decorator==4.4.2==py_0
expat==2.2.6==he6710b0_0
ffmpeg==4.0.2==ha0c5888_2
fontconfig==2.13.0==h9420a91_0
freeglut==3.0.0==hf484d3e_1005
freetype==2.9.1==h8a8886c_1
gast==0.2.2==py37_0
glib==2.63.1==h3eb4bd4_1
gmp==6.1.2==hf484d3e_1000
gnutls==3.5.19==h2a4e5f8_1
google-auth==1.14.1==py_0
google-auth-oauthlib==0.4.1==py_2
google-pasta==0.2.0==py_0
graphite2==1.3.13==he1b5a44_1001
grpcio==1.27.2==py37hf8bcb03_0
gst-plugins-base==1.14.0==hbbd80ab_1
gstreamer==1.14.0==hb31296c_0
h5py==2.8.0==py37h3010b51_1003
harfbuzz==1.8.8==hffaf4a1_0
hdf5==1.10.2==hc401514_3
icu==58.2==he6710b0_3
idna==2.9==py_1
imageio==2.8.0==py_0
intel-openmp==2020.1==217
jasper==2.0.14==h07fcdf6_1
jpeg==9b==h024ee3a_2
keras-applications==1.0.8==py_0
keras-base==2.3.1==py37_0
keras-gpu==2.3.1==0
keras-preprocessing==1.1.0==py_1
kiwisolver==1.2.0==py37hfd86e86_0
ld_impl_linux-64==2.33.1==h53a641e_7
libedit==3.1.20181209==hc058e9b_0
libffi==3.3==he6710b0_1
libgcc-ng==9.1.0==hdf63c60_0
libgfortran==3.0.0==1
libgfortran-ng==7.3.0==hdf63c60_0
libglu==9.0.0==he1b5a44_1001
libiconv==1.15==h516909a_1006
libopencv==3.4.2==hb342d67_1
libpng==1.6.37==hbc83047_0
libprotobuf==3.11.4==hd408876_0
libstdcxx-ng==9.1.0==hdf63c60_0
libtiff==4.1.0==h2733197_0
libuuid==1.0.3==h1bed415_2
libxcb==1.13==h1bed415_1
libxml2==2.9.9==hea5a465_1
markdown==3.1.1==py37_0
matplotlib==3.1.3==py37_0
matplotlib-base==3.1.3==py37hef1b27d_0
mkl==2020.1==217
mkl-service==2.3.0==py37he904b0f_0
mkl_fft==1.0.15==py37ha843d7b_0
mkl_random==1.1.0==py37hd6b4f25_0
ncurses==6.2==he6710b0_1
nettle==3.3==0
networkx==2.4==py_0
numpy==1.18.1==py37h4f9e942_0
numpy-base==1.18.1==py37hde5b4d6_1
oauthlib==3.1.0==py_0
olefile==0.46==py37_0
opencv==3.4.2==py37h6fd60c2_1
openh264==1.8.0==hdbcaa40_1000
openssl==1.1.1g==h516909a_0
opt_einsum==3.1.0==py_0
pcre==8.43==he6710b0_0
pillow==7.1.2==py37hb39fc2d_0
pip==20.0.2==py37_3
pixman==0.38.0==h516909a_1003
protobuf==3.11.4==py37he6710b0_0
py-opencv==3.4.2==py37hb342d67_1
pyasn1==0.4.8==py_0
pyasn1-modules==0.2.7==py_0
pycparser==2.20==py_0
pyjwt==1.7.1==py37_0
pyopenssl==19.1.0==py37_0
pyparsing==2.4.7==py_0
pyqt==5.9.2==py37h05f1152_2
pysocks==1.7.1==py37_0
python==3.7.7==hcff3b4d_5
python-dateutil==2.8.1==py_0
python_abi==3.7==1_cp37m
pywavelets==1.1.1==py37h7b6447c_0
pyyaml==5.3.1==py37h7b6447c_0
qt==5.9.7==h5867ecd_1
readline==8.0==h7b6447c_0
requests==2.23.0==py37_0
requests-oauthlib==1.3.0==py_0
rsa==4.0==py_0
scikit-image==0.16.2==py37h0573a6f_0
scipy==1.4.1==py37h0b6359f_0
setuptools==46.4.0==py37_0
sip==4.19.8==py37hf484d3e_0
six==1.14.0==py37_0
sqlite==3.31.1==h62c20be_1
tensorboard==2.1.0==py3_0
tensorflow==2.1.0==gpu_py37h7a4bb67_0
tensorflow-base==2.1.0==gpu_py37h6c5654b_0
tensorflow-estimator==2.1.0==pyhd54b08b_0
tensorflow-gpu==2.1.0==h0d30ee6_0
termcolor==1.1.0==py37_1
tk==8.6.8==hbc83047_0
toolz==0.10.0==py_0
tornado==6.0.4==py37h7b6447c_1
tqdm==4.46.0==py_0
urllib3==1.25.8==py37_0
werkzeug==1.0.1==py_0
wheel==0.34.2==py37_0
wrapt==1.12.1==py37h7b6447c_1
x264==1!152.20180806==h14c3975_0
xorg-fixesproto==5.0==h14c3975_1002
xorg-inputproto==2.3.2==h14c3975_1002
xorg-kbproto==1.0.7==h14c3975_1002
xorg-libx11==1.6.9==h516909a_0
xorg-libxau==1.0.9==h14c3975_0
xorg-libxext==1.3.4==h516909a_0
xorg-libxfixes==5.0.3==h516909a_1004
xorg-libxi==1.7.10==h516909a_0
xorg-xextproto==7.3.0==h14c3975_1002
xorg-xproto==7.0.31==h14c3975_1007
xz==5.2.5==h7b6447c_0
yaml==0.1.7==had09818_2
zlib==1.2.11==h7b6447c_3
zstd==1.3.7==h0b5b093_0